[move] #6 SelectableTheatersInModal 파일을 만들고 이곳으로 SelectableTheatersInModal, DetailRegionTheaters, TheaterListItem 함수들 이동
