+++ app/src/main/java/org/sopt/cgv/feature/time/TimeScreen.kt
@@ -8,7 +8,6 @@ import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.ExperimentalLayoutApi
 import androidx.compose.foundation.layout.FlowRow
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -20,15 +19,12 @@ import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Button
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
-import androidx.compose.material3.ModalBottomSheet
 import androidx.compose.material3.SheetState
 import androidx.compose.material3.Tab
 import androidx.compose.material3.TabRow
 import androidx.compose.material3.TabRowDefaults
 import androidx.compose.material3.TabRowDefaults.tabIndicatorOffset
 import androidx.compose.material3.Text
-import androidx.compose.material3.rememberModalBottomSheetState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.MutableState
 import androidx.compose.runtime.getValue
@@ -36,10 +32,8 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.drawBehind
-import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.geometry.Size
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.Paint
@@ -47,13 +41,10 @@ import androidx.compose.ui.graphics.RectangleShape
 import androidx.compose.ui.graphics.Shape
 import androidx.compose.ui.graphics.drawOutline
 import androidx.compose.ui.graphics.drawscope.drawIntoCanvas
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import kotlinx.coroutines.launch
-import org.sopt.cgv.R
 import org.sopt.cgv.core.designsystem.theme.CGVTheme
 
 @OptIn(ExperimentalMaterial3Api::class)
@@ -77,105 +68,11 @@ fun TimeScreen(
     )
 }
 
-@Composable
-fun SelectableTheatersInModal(
-    modifier: Modifier,
-    movieTheatersByDetailRegion: List<MovieTheatersByDetailRegion>
-) {
-    LazyColumn(
-        modifier = modifier
-            .padding(end = 18.dp)
-    ) {
-        item { Spacer(modifier = Modifier.height(16.dp)) }
-
-        item {
-            Text(
-                text = stringResource(R.string.cgv_theater_selection_guide),
-                style = CGVTheme.typography.small1_l_10
-            )
-        }
-
-        item { Spacer(modifier = Modifier.height(25.dp)) }
-
-        items(movieTheatersByDetailRegion) { movieTheatersByDetailRegion ->
-            DetailRegionTheaters(
-                theaterNames = movieTheatersByDetailRegion.theaterNames,
-                detailRegionName = movieTheatersByDetailRegion.detailRegionName
-            )
-        }
-    }
-}
 
 data class MovieTheatersByDetailRegion(
     val detailRegionName: String,
     val theaterNames: List<String>
 )
-
-@Composable
-fun DetailRegionTheaters(
-    theaterNames: List<String>,
-    detailRegionName: String
-) {
-    Column(
-        modifier = Modifier.fillMaxWidth()
-    ) {
-        Text(
-            text = detailRegionName,
-            style = CGVTheme.typography.head0_b_10,
-            color = Color.Red
-        )
-        Spacer(modifier = Modifier.height(25.dp))
-        theaterNames.forEach {
-            TheaterListItem(it)
-        }
-        Spacer(modifier = Modifier.height(36.dp))
-    }
-}
-
-@Composable
-fun TheaterListItem(
-    theaterName: String
-) {
-    var isSelected by remember { mutableStateOf(false) }
-
-    Box(
-        modifier = Modifier
-            .fillMaxWidth()
-            .clickable { isSelected = !isSelected }
-            .background(color = if (isSelected) Color.Gray else Color.White)
-            .drawBehind {
-                val strokeWidth = 1.dp.toPx()
-                val y = size.height - strokeWidth / 2
-                drawLine(
-                    color = Color(0xFFEDEDED),
-                    start = Offset(0f, y),
-                    end = Offset(size.width, y),
-                    strokeWidth = strokeWidth
-                )
-            }
-            .padding(horizontal = 8.dp, vertical = 11.dp)
-    ) {
-        Row(
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            Text(
-                text = theaterName,
-                style = CGVTheme.typography.body3_m_14
-            )
-
-            Spacer(modifier = Modifier.weight(1f))
-
-            if (isSelected) {
-                Icon(
-                    painter = painterResource(R.drawable.ic_time_modal_check),
-                    contentDescription = "",
-                    modifier = Modifier
-                        .padding(horizontal = 5.dp, vertical = 7.dp)
-                )
-            }
-        }
-    }
-}
 
 
 @Composable
